third_party/blink/web_tests/virtual/legacy-client-hints-no-fp-delegation/external/wpt/client-hints/accept-ch-stickiness/cross-origin-subresource-with-feature-policy.https-expected.txt

This is a testharness.js-based test.
PASS cross origin subresources authorized by FP gets it own resources precondition: Test that the browser does not have client hints preferences cached
PASS cross origin subresources authorized by FP gets it own resources set Accept-CH
FAIL cross origin subresources authorized by FP gets it own resources got client hints according to expectations. assert_equals: message from opened page expected "PASS" but got "FAIL"
Harness: the test ran to completion.
